백준 알고리즘 풀이(1000~1002) README 수정
