+++ README.md
@@ -1,1 +1,1 @@
-안녕하세요
+개인 알고리즘 공부 중입니다.
